move user functions out of config
